+++ templates/home.html
@@ -75,9 +75,12 @@
                                         <h5 class="text-white">Input exercises </h5>
                                         <div class="dropdown mx-3">
                                             <select id="exercise_type" name="exercise_type">
-                                                <option value="cardio">Cardio</option>
-                                                <option value="weightlifting">Lifting Weights</option>
-                                                <option value="swimming">Swimming</option>
+                                                <option value="Cardio">Cardio</option>
+                                                <option value="Swimming">Swimming</option>
+                                                <option value="Weightlifting">Lifting Weights</option>
+                                                <option value="Calisthenics">Calisthenics</option>
+                                                <option value="Yoga">Yoga</option>
+                                                <option value="Walking">Walking</option>
                                             </select>
                                         </div>
